test(spark): one backend-agnostic spark fixture + the P0 real-Spark lane
Executes P0 of docs/superpowers/specs/2026-08-10-spark-native-execution-design.md
(plan: docs/superpowers/plans/2026-08-10-p0-real-spark-connect.md). Tests and CI
only; NO engine code changed.

THE OBSTACLE: the `spark` fixture was duplicated across NINE test files, each
hardcoding `pysail.spark.SparkConnectServer`. The tier had never been run against
real Spark because the tests could not express it.

FINDING THE PLAN MISSED (and it would have made the new lane a lie): these files
also carried MODULE-LEVEL `pytest.importorskip("pysail")`. On a lane where pysail
is deliberately absent, every test would SKIP and the lane would report green
while proving nothing -- the same shape as the zero-polars gate trap, one level
down. The pysail guard is now the fixture's business, not the module's; each file
keeps a `pyspark` guard instead.

DEVIATION FROM THE PLAN, on evidence: the plan put the fixture in a separate
tests/conftest_spark.py imported by each file, to avoid the root conftest growing
a pyspark import. That import pattern produced 41 ruff errors -- F811 in every
file whose tests take `spark` as a parameter. The concern behind it is void: the
root conftest holds only stdlib + pytest, and this fixture's imports are ALL
inside the body, so a suite that never requests it pays nothing at collection.
Fixture now lives in tests/conftest.py; no cross-file imports, no noqa noise.

Two files (identity_parity, identity_incremental) kept their guards INSIDE the
fixture rather than at module level; an assertion caught that mid-run rather than
silently producing files with no pyspark guard at all. Their stale docstrings are
corrected too.

test_sail_connectivity.py built its server inline in the test body rather than a
fixture; rewritten to use the shared one, so the S1 harness gate now proves
whichever backend the env selects.

THE LANE: installs pyspark[connect]>=4 and NOT pysail, asserts pysail is absent
(a lane that silently picked it up from the lockfile would prove nothing), and
runs the same tests with GOLDENMATCH_SPARK_REMOTE=local[*] -- Spark's own
in-process Connect server, no container, no cluster. goldenmatch[sail] pins
pyspark<4 because of pysail, so the client is installed directly.

Deliberately continue-on-error and NOT in ci-required: while P0 runs, a red is a
finding to enumerate, not a merge blocker. Verified: spark_connect present,
continue-on-error true, absent from ci-required.

Local verification is collection + compile + ruff only. Running it needs pyspark
and a JVM, which OOMs the dev box -- CI is the harness.
